+++ AlleNeune/Fine.swift
@@ -9,7 +9,7 @@ struct Fine {
     //attributes
     var id : Int?
     var name : String?
-    var amount : Int?
+    var amount : Double?
     
     //api parameters
     static let ROOT = "fine"
@@ -24,7 +24,7 @@ struct Fine {
     static let GET_BY_CLUB : String = Fine.GENERIC_URL + "/get_by_club"
     static let UPDATE : String = Fine.GENERIC_URL + "/"
     
-    init(id: Int, name: String, amount: Int) {
+    init(id: Int, name: String, amount: Double) {
         self.id = id
         self.name = name
         self.amount = amount
